msm: thermal: simple: Remove inefficient code
A per-CPU definition is not necessary for throttle_policy, as we will enforce the same throttle policy on all CPUs. Also, there is no reason to forcefully queue everything on CPU0; instead, just ensure all workqueue items are non-reentrant on all CPUs (as we don't want the main thermal worker running concurrently on different CPUs).

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Oveno <ovenoboyo@gmail.com>