msm: thermal: simple: Fix lack of throttling when temp suddenly jumps
It's possible for the temperature to start off very high at boot and be
greater than zone1's trip temperature, causing the thermal driver to do
nothing when throttling is needed most.

It's also possible for the temperature to jump significantly in between
polling cycles and trigger this bug.

Fix this by making sure that the temperature is indeed below zone0's trip
point before declaring that the CPU does not need to be throttled.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Raphiel Rollerscaperers <raphielscape@outlook.com>